Task/options file for commands (#18)
* Get command oriented options file names

* Change from Command to string

* Remove superfluous include

* Replace static calls with ParserContext

* Context is a property of Arguments

* Method renames

* Use context in UsageDisplay

* Context files

* App Commands now all have a context

* stash

* Tests fix

* Warning

* DNXApp code moved to cpp files

* More file organising

* Docs tidy up

* more doc tweaks

+++ DNX.Utils/ListUtils.h
@@ -53,7 +53,7 @@ namespace DNX::Utils
         template<class T>
         static T& GetRandom(list<T>& list)
         {
-            const auto index = MathUtils::GetRandomNumber(0, list.size(), false);
+            const auto index = MathUtils::GetRandomNumber(0, static_cast<int>(list.size()), false);
 
             T& result = GetAt(list, index);
